Refactor EvoCriteria to be a struct, reduce allocation (#3483)
* Make EvolutionCriteria struct

8 bytes per object instead of 26
Unify LevelMin/LevelMax to match EncounterTemplate
bubble up precise array type for better iteration

* Inline queue operations, less allocation

* Inline some logic

* Update EvolutionChain.cs

* Improve clarity on duplicate move check

* Search reverse

For a dual stage chain, finds it first iteration rather than second.

+++ PKHeX.Core/Legality/Evolutions/EvolutionSets/EvolutionSet3.cs
@@ -25,7 +25,7 @@ namespace PKHeX.Core
                 case 6: /* Trade while holding */
                     return new EvolutionMethod(method, species, argument: arg);
                 case 4: /* Level Up */
-                    return new EvolutionMethod(4, species, argument: arg, level:arg);
+                    return new EvolutionMethod(4, species, argument: arg, level: (byte)arg);
                 case 7: /* Use item */
                 case 15: /* Beauty evolution*/
                     return new EvolutionMethod(method + 1, species, argument: arg);
@@ -36,7 +36,7 @@ namespace PKHeX.Core
                 case 12: /* Wurmple -> Cascoon evolution */
                 case 13: /* Nincada -> Ninjask evolution */
                 case 14: /* Shedinja spawn in Nincada -> Ninjask evolution */
-                    return new EvolutionMethod(method + 1, species, argument: arg, level: arg);
+                    return new EvolutionMethod(method + 1, species, argument: arg, level: (byte)arg);
 
                 default:
                     throw new ArgumentOutOfRangeException(nameof(method));
